Update tests now that failed user configured rpc endpoint health raises exception. Update test fixtures for endpoints and health checks.
