+++ tests/acceptance/conftest.py
@@ -1,10 +1,9 @@
 import random
 
 import pytest
-from web3 import HTTPProvider, Web3
+from web3 import Web3
 
 import tests
-from nucypher.blockchain.eth.actors import Operator
 from nucypher.blockchain.eth.agents import (
     ContractAgency,
     CoordinatorAgent,
@@ -16,7 +15,6 @@ from nucypher.blockchain.eth.constants import NULL_ADDRESS
 from nucypher.blockchain.eth.interfaces import BlockchainInterfaceFactory
 from nucypher.blockchain.eth.registry import ContractRegistry, RegistrySourceManager
 from nucypher.crypto.powers import TransactingPower
-from nucypher.utilities.endpoint import RPCEndpoint
 from nucypher.utilities.logging import Logger
 from tests.constants import (
     BONUS_TOKENS_FOR_TESTS,
@@ -577,25 +575,6 @@ def mock_condition_blockchains(module_mocker):
 
     module_mocker.patch(
         "nucypher.blockchain.eth.domains.get_domain", return_value=TEMPORARY_DOMAIN
-    )
-
-
-@pytest.fixture(scope="module", autouse=True)
-def mock_multichain_configuration(module_mocker, testerchain):
-    module_mocker.patch.object(
-        Operator, "_make_condition_provider", return_value=testerchain.provider
-    )
-
-    def _mock_make_provider(endpoint, session, request_kwargs):
-        if endpoint == TEST_ETH_PROVIDER_URI:
-            return testerchain.provider
-        else:
-            return HTTPProvider(
-                endpoint, session=session, request_kwargs=request_kwargs
-            )
-
-    module_mocker.patch.object(
-        RPCEndpoint, "_make_provider", side_effect=_mock_make_provider
     )
